refactor(admin): finish stage 4 app organization
